+++ tests/daemon_integration.rs
@@ -1,3 +1,5 @@
+#![cfg(feature = "slow-tests")]
+
 use assert_cmd::cargo::cargo_bin;
 use serde_json::Value;
 use std::path::PathBuf;